Implemented layer sampling

+++ True Seamless Texturing/Assets/Repetitionless/Shaders/HLSL/SampleSeamlessMaterialTerrain.hlsl
@@ -11,7 +11,7 @@ void SampleSeamlessMaterialTerrain_float(
     SamplerState SS, float2 UV, float3 WorldNormalVector, // Control requires clamped sampler state, otherwise fades off at edges of terrain
     float3 WorldPosition, float3 CameraPosition, // Positions
     UnityTexture2D Holes, UnityTexture2D Control,
-    float SurfaceType, float DebuggingIndex, // Enums
+    int SurfaceType, int DebuggingIndex, // Enums
 
     // ------- TERRAIN LAYER 1 ------- //
